876: Moved actions to separate files to prevent circular dependency errors
