+++ src/components/Page/Page.tsx
@@ -8,8 +8,8 @@ import { useWeb3React } from "@web3-react/core";
 import { useAppDispatch } from "../../app/hooks";
 import { WalletProvider } from "../../constants/supportedWalletProviders";
 import { InterfaceContext } from "../../contexts/interface/Interface";
-import { resetOrders } from "../../features/orders/ordersSlice";
-import { useTransactions } from "../../features/transactions/transactionHooks";
+import { resetOrders } from "../../features/orders/ordersActions";
+import { useTransactions } from "../../features/transactions/transactionsHooks";
 import useHistoricalTransactions from "../../features/transactions/useHistoricalTransactions";
 import useTransactionsFilterFromLocalStorage from "../../features/transactions/useTransactionsFilterFromLocalStorage";
 import { Wallet } from "../../features/wallet/Wallet";
