+++ src/contexts/lastLook/LastLook.tsx
@@ -16,7 +16,15 @@ import { useAppDispatch, useAppSelector } from "../../app/hooks";
 import { notifyError } from "../../components/Toasts/ToastController";
 import { LAST_LOOK_ORDER_EXPIRY_SEC } from "../../constants/configParams";
 import { SubmittedTransactionWithOrder } from "../../entities/SubmittedTransaction/SubmittedTransaction";
-import { selectProtocolFee } from "../../features/metadata/metadataSlice";
+// import { submitTransactionWithExpiry } from "../../features/transactions/transactionsSlice";
+import {
+  transformToSubmittedLastLookOrder,
+  transformToSubmittedRFQOrder,
+} from "../../entities/SubmittedTransaction/SubmittedTransactionTransformers";
+import {
+  selectAllTokenInfo,
+  selectProtocolFee,
+} from "../../features/metadata/metadataSlice";
 import { updatePricing } from "../../features/pricing/pricingSlice";
 import { TradeTerms } from "../../features/tradeTerms/tradeTermsSlice";
 import { submitTransactionWithExpiry } from "../../features/transactions/transactionsSlice";
@@ -56,6 +64,7 @@ export const LastLookContext = createContext<{
 const connectedServers: Record<string, Server> = {};
 const LastLookProvider: FC = ({ children }) => {
   const { account, library, chainId } = useWeb3React();
+  const tokens = useAppSelector(selectAllTokenInfo);
 
   const { t } = useTranslation();
 
@@ -151,12 +160,14 @@ const LastLookProvider: FC = ({ children }) => {
         signerAmount: isSell ? baseAmountAtomic : quoteAmountAtomic,
         senderAmount: !isSell ? baseAmountAtomic : quoteAmountAtomic,
       });
+
       const signature = await createOrderERC20Signature(
         unsignedOrder,
         library.getSigner(),
         SwapERC20.getAddress(chainId!) || "",
         chainId!
       );
+
       const order: OrderERC20 = {
         expiry: unsignedOrder.expiry,
         nonce: unsignedOrder.nonce,
@@ -168,15 +179,15 @@ const LastLookProvider: FC = ({ children }) => {
         ...signature,
       };
 
-      const transaction: SubmittedTransactionWithOrder = {
-        type: "Order",
-        order: order,
-        nonce: order.nonce,
-        status: "processing",
-        protocol: "last-look-erc20",
-        expiry: unsignedOrder.expiry,
-        timestamp: Date.now(),
-      };
+      const signerToken = tokens.find((t) => t.address === order.signerToken);
+      const senderToken = tokens.find((t) => t.address === order.senderToken);
+
+      const transaction = transformToSubmittedLastLookOrder(
+        order,
+        signerToken!,
+        senderToken!
+      );
+
       dispatch(
         submitTransactionWithExpiry({
           transaction,
@@ -203,6 +214,7 @@ const LastLookProvider: FC = ({ children }) => {
       const { locator, order } = params;
       const server = connectedServers[locator];
       try {
+        console.log(order);
         return server.considerOrderERC20(order);
       } catch (e) {
         console.error("Server unable to consider order: ", e);
