+++ src/features/orders/ordersSlice.ts
@@ -1,64 +1,20 @@
-import { Server } from "@airswap/libraries";
-import {
-  toAtomicString,
-  FullOrderERC20,
-  OrderERC20,
-  Levels,
-  TokenInfo,
-} from "@airswap/utils";
-import {
-  createAsyncThunk,
-  createSelector,
-  createSlice,
-  Dispatch,
-  PayloadAction,
-} from "@reduxjs/toolkit";
+import { OrderERC20, Levels } from "@airswap/utils";
+import { createSelector, createSlice, PayloadAction } from "@reduxjs/toolkit";
 
 import BigNumber from "bignumber.js";
-import { providers } from "ethers";
 
-import { AppDispatch, RootState } from "../../app/store";
-import { notifyRejectedByUserError } from "../../components/Toasts/ToastController";
-import {
-  RFQ_EXPIRY_BUFFER_MS,
-  RFQ_MINIMUM_REREQUEST_DELAY_MS,
-} from "../../constants/configParams";
-import nativeCurrency from "../../constants/nativeCurrency";
-import {
-  SubmittedRFQOrder,
-  SubmittedWithdrawTransaction,
-} from "../../entities/SubmittedTransaction/SubmittedTransaction";
-import {
-  transformToSubmittedApprovalTransaction,
-  transformToSubmittedDepositTransaction,
-  transformToSubmittedWithdrawTransaction,
-} from "../../entities/SubmittedTransaction/SubmittedTransactionTransformers";
-import { AppError, AppErrorType, isAppError } from "../../errors/appError";
-import transformUnknownErrorToAppError from "../../errors/transformUnknownErrorToAppError";
-import getWethAddress from "../../helpers/getWethAddress";
-import toRoundedAtomicString from "../../helpers/toRoundedAtomicString";
+import { RootState } from "../../app/store";
+import { AppError } from "../../errors/appError";
 import { gasUsedPerSwap } from "../gasCost/gasCostApi";
 import { selectGasPriceInQuoteTokens } from "../gasCost/gasCostSlice";
 import { selectBestPricing } from "../pricing/pricingSlice";
 import { selectTradeTerms } from "../tradeTerms/tradeTermsSlice";
 import {
-  declineTransaction,
-  revertTransaction,
-  submitTransaction,
-} from "../transactions/transactionActions";
-import { submitTransactionWithExpiry } from "../transactions/transactionsSlice";
-import {
   setWalletConnected,
   setWalletDisconnected,
 } from "../wallet/walletSlice";
-import {
-  approveToken,
-  depositETH,
-  orderSortingFunction,
-  requestOrders,
-  takeOrder,
-  withdrawETH,
-} from "./ordersApi";
+import { approve, deposit, request, take } from "./ordersActions";
+import { orderSortingFunction } from "./ordersApi";
 
 export interface OrdersState {
   orders: OrderERC20[];
@@ -73,289 +29,6 @@ const initialState: OrdersState = {
   reRequestTimerId: null,
   errors: [],
 };
-
-// replaces WETH to ETH on Wrapper orders
-const refactorOrder = (order: OrderERC20, chainId: number) => {
-  let newOrder = { ...order };
-  if (order.senderToken === getWethAddress(chainId)) {
-    newOrder.senderToken = nativeCurrency[chainId].address;
-  } else if (order.signerToken === getWethAddress(chainId)) {
-    newOrder.signerToken = nativeCurrency[chainId].address;
-  }
-  return newOrder;
-};
-
-export const handleOrderError = (dispatch: Dispatch, error: any) => {
-  const appError = transformUnknownErrorToAppError(error);
-
-  if (appError.error && "message" in appError.error) {
-    dispatch(declineTransaction(appError.error.message));
-  }
-
-  if (appError.type === AppErrorType.rejectedByUser) {
-    notifyRejectedByUserError();
-  } else {
-    dispatch(setErrors([appError]));
-  }
-};
-
-export const deposit = createAsyncThunk(
-  "orders/deposit",
-  async (
-    params: {
-      chainId: number;
-      senderAmount: string;
-      senderToken: TokenInfo;
-      provider: providers.Web3Provider;
-    },
-    { dispatch }
-  ) => {
-    try {
-      const tx = await depositETH(
-        params.chainId,
-        params.senderAmount,
-        params.senderToken.decimals,
-        params.provider
-      );
-
-      if (!tx.hash) {
-        console.error("Approval transaction hash is missing.");
-
-        return;
-      }
-
-      const amount = toAtomicString(
-        params.senderAmount,
-        params.senderToken.decimals
-      );
-
-      const transaction = transformToSubmittedDepositTransaction(
-        tx.hash,
-        params.senderToken,
-        nativeCurrency[params.chainId],
-        amount
-      );
-
-      dispatch(submitTransaction(transaction));
-    } catch (e: any) {
-      handleOrderError(dispatch, e);
-      throw e;
-    }
-  }
-);
-
-export const resetOrders = createAsyncThunk(
-  "orders/reset",
-  async (params: undefined, { getState, dispatch }) => {
-    await dispatch(setResetStatus());
-    dispatch(clear());
-  }
-);
-
-export const withdraw = createAsyncThunk(
-  "orders/withdraw",
-  async (
-    params: {
-      chainId: number;
-      senderAmount: string;
-      senderToken: TokenInfo;
-      provider: any;
-    },
-    { getState, dispatch }
-  ) => {
-    try {
-      const tx = await withdrawETH(
-        params.chainId,
-        params.senderAmount,
-        params.senderToken.decimals,
-        params.provider
-      );
-
-      if (!tx.hash) {
-        console.error("Approval transaction hash is missing.");
-
-        return;
-      }
-
-      const amount = toAtomicString(
-        params.senderAmount,
-        params.senderToken.decimals
-      );
-
-      const transaction = transformToSubmittedWithdrawTransaction(
-        tx.hash,
-        nativeCurrency[params.chainId],
-        params.senderToken,
-        amount
-      );
-
-      dispatch(submitTransaction(transaction));
-    } catch (e: any) {
-      handleOrderError(dispatch, e);
-      throw e;
-    }
-  }
-);
-
-export const request = createAsyncThunk(
-  "orders/request",
-  async (
-    params: {
-      servers: Server[];
-      signerToken: string;
-      senderToken: string;
-      senderAmount: string;
-      senderTokenDecimals: number;
-      senderWallet: string;
-      proxyingFor?: string;
-    },
-    { dispatch }
-  ) => {
-    const orders = await requestOrders(
-      params.servers,
-      params.signerToken,
-      params.senderToken,
-      params.senderAmount,
-      params.senderTokenDecimals,
-      params.senderWallet,
-      params.proxyingFor
-    );
-    if (orders.length) {
-      const bestOrder = [...orders].sort(orderSortingFunction)[0];
-      const now = Date.now();
-      const expiry = parseInt(bestOrder.expiry) * 1000;
-      // Due to the sorting in orderSorting function, these orders will be at
-      // the bottom of the list, so if the best one has a very short expiry
-      // so do all the others. Return an empty order array as none are viable.
-      if (expiry - now < RFQ_EXPIRY_BUFFER_MS) return [];
-
-      const timeTilReRequest = Math.max(
-        expiry - now - RFQ_EXPIRY_BUFFER_MS,
-        RFQ_MINIMUM_REREQUEST_DELAY_MS
-      );
-      const reRequestTimerId = window.setTimeout(
-        () => dispatch(request(params)),
-        timeTilReRequest
-      );
-      dispatch(setReRequestTimerId(reRequestTimerId));
-    }
-    return orders;
-  }
-);
-
-export const approve = createAsyncThunk<
-  // Return type of the payload creator
-  void,
-  // Params
-  {
-    token: TokenInfo;
-    library: any;
-    contractType: "Wrapper" | "Swap";
-    chainId: number;
-    amount: string;
-  },
-  // Types for ThunkAPI
-  {
-    // thunkApi
-    dispatch: AppDispatch;
-    state: RootState;
-  }
->("orders/approve", async (params, { getState, dispatch }) => {
-  try {
-    const { token, contractType, amount } = params;
-    const approveAmount = toRoundedAtomicString(amount, token.decimals);
-
-    const tx = await approveToken(
-      token.address,
-      params.library,
-      contractType,
-      approveAmount
-    );
-
-    if (!tx.hash) {
-      console.error("Approval transaction hash is missing.");
-
-      return;
-    }
-
-    const transaction = transformToSubmittedApprovalTransaction(
-      tx.hash,
-      token,
-      approveAmount
-    );
-    dispatch(submitTransaction(transaction));
-  } catch (e: any) {
-    handleOrderError(dispatch, e);
-    throw e;
-  }
-});
-
-export const take = createAsyncThunk<
-  // Return type of the payload creator
-  void,
-  // Params
-  {
-    order: OrderERC20 | FullOrderERC20;
-    library: any;
-    contractType: "Swap" | "Wrapper";
-    onExpired: () => void;
-  },
-  // Types for ThunkAPI
-  {
-    dispatch: AppDispatch;
-    state: RootState;
-  }
->("orders/take", async (params, { getState, dispatch }) => {
-  const tx = await takeOrder(params.order, params.library, params.contractType);
-
-  if (isAppError(tx)) {
-    const appError = tx;
-    if (appError.type === AppErrorType.rejectedByUser) {
-      notifyRejectedByUserError();
-      dispatch(
-        revertTransaction({
-          signerWallet: params.order.signerWallet,
-          nonce: params.order.nonce,
-          reason: appError.type,
-        })
-      );
-    } else {
-      dispatch(setErrors([appError]));
-    }
-
-    if (appError.error && "message" in appError.error) {
-      dispatch(declineTransaction(appError.error.message));
-    }
-
-    throw appError;
-  }
-
-  // When dealing with the Wrapper, since the "actual" swap is ETH <-> ERC20,
-  // we should change the order tokens to WETH -> ETH
-  let newOrder =
-    params.contractType === "Swap"
-      ? params.order
-      : refactorOrder(params.order, params.library._network.chainId);
-  if (tx.hash) {
-    const transaction: SubmittedRFQOrder = {
-      type: "Order",
-      order: newOrder,
-      protocol: "request-for-quote-erc20",
-      hash: tx.hash,
-      status: "processing",
-      timestamp: Date.now(),
-      nonce: params.order.nonce,
-      expiry: params.order.expiry,
-    };
-    dispatch(
-      submitTransactionWithExpiry({
-        transaction,
-        signerWallet: params.order.signerWallet,
-        onExpired: params.onExpired,
-      })
-    );
-  }
-});
 
 export const ordersSlice = createSlice({
   name: "orders",
