+++ src/features/registry/registrySlice.ts
@@ -1,15 +1,11 @@
-import { createAsyncThunk, createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
-import { providers } from "ethers";
-import uniqBy from "lodash.uniqby";
-
-import { AppDispatch, RootState } from "../../app/store";
-import { getActiveTokensFromLocalStorage } from "../metadata/metadataApi";
+import { RootState } from "../../app/store";
 import {
   setWalletConnected,
   setWalletDisconnected,
 } from "../wallet/walletSlice";
-import { getStakerTokens } from "./registryApi";
+import { fetchSupportedTokens } from "./registryActions";
 
 export interface RegistryState {
   stakerTokens: Record<string, string[]>;
@@ -22,39 +18,6 @@ const initialState: RegistryState = {
   allSupportedTokens: [],
   status: "idle",
 };
-
-export const fetchSupportedTokens = createAsyncThunk<
-  {
-    allSupportedTokens: string[];
-    stakerTokens: Record<string, string[]>;
-    activeTokens: string[];
-  },
-  {
-    provider: providers.Provider;
-  },
-  {
-    // Optional fields for defining thunkApi field types
-    dispatch: AppDispatch;
-    state: RootState;
-  }
->("registry/fetchSupportedTokens", async ({ provider }, { getState }) => {
-  const { wallet } = getState();
-  const stakerTokens = await getStakerTokens(wallet.chainId!, provider);
-  // Combine token lists from all makers and flatten them.
-  const allSupportedTokens = uniqBy(
-    Object.values(stakerTokens).flat(),
-    (i) => i
-  );
-  const activeTokensLocalStorage = getActiveTokensFromLocalStorage(
-    wallet.address!,
-    wallet.chainId!
-  );
-  const activeTokens =
-    (activeTokensLocalStorage.length && activeTokensLocalStorage) ||
-    allSupportedTokens ||
-    [];
-  return { stakerTokens, allSupportedTokens, activeTokens };
-});
 
 export const registrySlice = createSlice({
   name: "registry",
