+++ src/features/transactions/useTransactionsFilterFromLocalStorage.ts
@@ -4,12 +4,12 @@ import { useLocalStorageValue } from "@react-hookz/web/esm";
 import { useWeb3React } from "@web3-react/core";
 
 import { useAppDispatch, useAppSelector } from "../../app/hooks";
-import { getTransactionsFilterLocalStorageKey } from "./transactionUtils";
 import {
   selectTransactionsFilter,
   setFilters,
   TransactionsState,
 } from "./transactionsSlice";
+import { getTransactionsFilterLocalStorageKey } from "./transactionsUtils";
 
 const useTransactionsFilterFromLocalStorage = () => {
   const dispatch = useAppDispatch();
