+++ src/features/transactions/useTransactionsFromLocalStorage.ts
@@ -2,7 +2,7 @@ import { useLocalStorageValue } from "@react-hookz/web/esm";
 import { useWeb3React } from "@web3-react/core";
 
 import { SubmittedTransaction } from "../../entities/SubmittedTransaction/SubmittedTransaction";
-import { getTransactionsLocalStorageKey } from "./transactionUtils";
+import { getTransactionsLocalStorageKey } from "./transactionsUtils";
 
 const useTransactionsFromLocalStorage = () => {
   const { account, chainId } = useWeb3React();
